refactor(profile): flattened profile file structure
