--- messages/profile/schema.py
@@ -1,6 +1,0 @@
-profile_schema = {
-  "TYPE": "PROFILE",
-  "USER_ID": {"type": str, "required": True},
-  "DISPLAY_NAME": {"type": str, "required": True},
-  "STATUS": {"type": str, "required": True},
-}
