+++ router.py
@@ -11,12 +11,11 @@
 # TODO: Implement complex flow (i.e. acks retries timeouts)
 import config
 import messages.utils.format as format
-import messages.profile.receiver as profile_receiver
-
+import messages.profile as profile
 def route(data, address):
   parsed_data = data.decode(config.ENCODING)
   message = format.deserialize_message(parsed_data)
   type = message.get("TYPE")
 
   if type == "PROFILE":
-    profile_receiver.receive(message)
+    profile.receive(message)
